missing_persons_ai/requirements.txt
streamlit>=1.32.0
pandas>=2.1.0
numpy>=1.26.0
pillow>=10.0.0
pyyaml>=6.0
opencv-python-headless
mediapipe>=0.10.0
scikit-learn>=1.3.0
sqlmodel>=0.0.14
greenlet==3.0.3
bcrypt>=4.0.0
jinja2
